MAINT: Move _SequenceReprBuilder to separate file
Now lives in skbio.sequence._repr.py
